Create lib for nimiq-rpc-client
